requirements.txt
greenlet==3.2.3
playwright==1.52.0
pyee==13.0.0
PyMuPDF==1.26.1
python-dotenv==1.1.1
typing_extensions==4.14.0
Unidecode==1.4.0
